@@ -8,114 +8,115 @@
     "node_modules"
   ],
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.17",
-    "@fortawesome/free-solid-svg-icons": "5.8.1",
-    "@fortawesome/react-fontawesome": "0.1.4",
+    "@fortawesome/fontawesome-svg-core": "1.2.26",
+    "@fortawesome/free-solid-svg-icons": "5.12.0",
+    "@fortawesome/react-fontawesome": "0.1.8",
     "availity-reactstrap-validation": "2.5.1",
     "axios": "0.19.0",
-    "bootstrap": "4.3.1",
+    "bootstrap": "4.4.1",
     "firebase": "^7.6.1",
     "loaders.css": "0.1.2",
-    "lodash": "4.17.13",
+    "lodash": "4.17.15",
     "moment": "2.24.0",
-    "react": "16.8.5",
-    "react-dom": "16.8.5",
-    "react-hot-loader": "4.8.0",
-    "react-jhipster": "0.9.0",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
+    "react-hot-loader": "4.12.18",
+    "react-jhipster": "0.9.1",
     "react-loadable": "5.5.0",
-    "react-redux": "6.0.1",
-    "react-redux-loading-bar": "4.2.0",
-    "react-router-dom": "5.0.0",
+    "react-redux": "7.1.3",
+    "react-redux-loading-bar": "4.5.0",
+    "react-router-dom": "5.1.2",
     "react-toastify": "4.5.2",
     "react-transition-group": "2.7.0",
     "reactstrap": "7.1.0",
-    "redux": "4.0.1",
+    "redux": "4.0.4",
     "redux-devtools": "3.5.0",
     "redux-devtools-dock-monitor": "1.1.3",
     "redux-devtools-log-monitor": "1.4.0",
-    "redux-promise-middleware": "6.1.0",
+    "redux-promise-middleware": "6.1.2",
     "redux-thunk": "2.3.0",
     "rxjs": "6.4.0",
     "sockjs-client": "1.3.0",
-    "tslib": "1.9.3",
-    "uuid": "3.3.2",
+    "tslib": "1.10.0",
+    "uuid": "3.3.3",
     "webstomp-client": "1.2.6"
   },
   "devDependencies": {
+    "@openapitools/openapi-generator-cli": "0.0.14-4.0.2",
     "@types/enzyme": "3.9.1",
-    "@types/jest": "24.0.11",
-    "@types/lodash": "4.14.123",
-    "@types/node": "10.14.3",
-    "@types/react": "16.8.10",
-    "@types/react-dom": "16.8.3",
+    "@types/jest": "24.0.23",
+    "@types/lodash": "4.14.149",
+    "@types/node": "12.12.17",
+    "@types/react": "16.9.16",
+    "@types/react-dom": "16.9.4",
     "@types/react-redux": "6.0.6",
-    "@types/react-router-dom": "4.3.1",
+    "@types/react-router-dom": "5.1.3",
     "@types/redux": "3.6.31",
-    "@types/webpack-env": "1.13.9",
-    "@typescript-eslint/eslint-plugin": "2.0.0",
-    "@typescript-eslint/parser": "2.0.0",
-    "autoprefixer": "9.5.0",
+    "@types/webpack-env": "1.14.1",
+    "@typescript-eslint/eslint-plugin": "2.11.0",
+    "@typescript-eslint/parser": "2.11.0",
+    "autoprefixer": "9.7.3",
     "base-href-webpack-plugin": "2.0.0",
-    "browser-sync": "2.26.5",
+    "browser-sync": "2.26.7",
     "browser-sync-webpack-plugin": "2.2.2",
-    "cache-loader": "2.0.1",
-    "copy-webpack-plugin": "5.0.2",
-    "core-js": "3.0.0",
-    "cross-env": "5.2.0",
-    "css-loader": "2.1.1",
-    "enzyme": "3.9.0",
-    "enzyme-adapter-react-16": "1.11.2",
-    "enzyme-to-json": "3.3.5",
-    "eslint": "6.1.0",
+    "cache-loader": "4.1.0",
+    "copy-webpack-plugin": "5.1.0",
+    "core-js": "3.4.8",
+    "cross-env": "6.0.3",
+    "css-loader": "3.3.0",
+    "enzyme": "3.10.0",
+    "enzyme-adapter-react-16": "^1.15.2",
+    "enzyme-to-json": "3.4.3",
+    "eslint": "6.7.2",
     "eslint-config-jhipster": "0.0.1",
     "eslint-config-jhipster-react": "0.0.1",
-    "eslint-config-prettier": "6.0.0",
-    "eslint-loader": "2.2.1",
-    "eslint-plugin-react": "7.14.3",
-    "file-loader": "3.0.1",
-    "fork-ts-checker-webpack-plugin": "1.5.0",
+    "eslint-config-prettier": "6.7.0",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-react": "7.17.0",
+    "file-loader": "5.0.2",
+    "fork-ts-checker-webpack-plugin": "3.1.1",
     "friendly-errors-webpack-plugin": "1.7.0",
-    "generator-jhipster": "6.4.1",
+    "generator-jhipster": "6.6.0",
     "html-webpack-plugin": "3.2.0",
-    "husky": "3.0.4",
+    "husky": "3.1.0",
     "identity-obj-proxy": "3.0.0",
-    "jest": "24.5.0",
-    "jest-junit": "8.0.0",
+    "jest": "24.9.0",
+    "jest-junit": "10.0.0",
     "jest-sonar-reporter": "2.0.0",
     "json-loader": "0.5.7",
-    "lint-staged": "8.2.1",
-    "merge-jsons-webpack-plugin": "1.0.18",
-    "mini-css-extract-plugin": "0.5.0",
-    "moment-locales-webpack-plugin": "1.0.7",
-    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "lint-staged": "9.5.0",
+    "merge-jsons-webpack-plugin": "1.0.20",
+    "mini-css-extract-plugin": "0.8.0",
+    "moment-locales-webpack-plugin": "1.1.2",
+    "optimize-css-assets-webpack-plugin": "5.0.3",
     "postcss-loader": "3.0.0",
-    "prettier": "1.18.2",
+    "prettier": "1.19.1",
     "react-infinite-scroller": "1.2.4",
-    "redux-mock-store": "1.5.3",
-    "rimraf": "2.6.3",
-    "sass": "1.17.3",
-    "sass-loader": "7.1.0",
+    "react-test-renderer": "^16.12.0",
+    "redux-mock-store": "1.5.4",
+    "rimraf": "3.0.0",
+    "sass": "1.23.7",
+    "sass-loader": "8.0.0",
     "simple-progress-webpack-plugin": "1.1.2",
-    "sinon": "7.3.0",
+    "sinon": "7.5.0",
     "source-map-loader": "0.2.4",
     "sourcemap-istanbul-instrumenter-loader": "0.2.0",
     "stripcomment-loader": "0.1.0",
-    "style-loader": "0.23.1",
-    "swagger-ui": "2.2.10",
-    "terser-webpack-plugin": "1.2.3",
-    "thread-loader": "2.1.2",
-    "to-string-loader": "1.1.5",
-    "ts-jest": "24.0.0",
-    "ts-loader": "5.3.3",
-    "typescript": "3.3.4000",
-    "@openapitools/openapi-generator-cli": "0.0.14-4.0.2",
-    "webpack": "4.28.4",
-    "webpack-cli": "3.3.0",
-    "webpack-dev-server": "3.2.1",
-    "webpack-merge": "4.2.1",
-    "webpack-notifier": "1.7.0",
-    "workbox-webpack-plugin": "4.1.1",
-    "write-file-webpack-plugin": "4.5.0"
+    "style-loader": "1.0.1",
+    "swagger-ui-dist": "3.24.3",
+    "terser-webpack-plugin": "2.2.3",
+    "thread-loader": "2.1.3",
+    "to-string-loader": "1.1.6",
+    "ts-jest": "24.2.0",
+    "ts-loader": "6.2.1",
+    "typescript": "3.7.3",
+    "webpack": "4.41.2",
+    "webpack-cli": "3.3.10",
+    "webpack-dev-server": "3.9.0",
+    "webpack-merge": "4.2.2",
+    "webpack-notifier": "1.8.0",
+    "workbox-webpack-plugin": "4.3.1",
+    "write-file-webpack-plugin": "4.5.1"
   },
   "engines": {
     "node": ">=8.9.0"
