@@ -1,6 +1,7 @@
 package com.riskbusters.norisknofun;
 
-import com.riskbusters.norisknofun.config.DefaultProfileUtil;
+import io.github.jhipster.config.DefaultProfileUtil;
+
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
