@@ -1,8 +1,10 @@
 package com.riskbusters.norisknofun;
 
 import com.riskbusters.norisknofun.config.ApplicationProperties;
-import com.riskbusters.norisknofun.config.DefaultProfileUtil;
+
+import io.github.jhipster.config.DefaultProfileUtil;
 import io.github.jhipster.config.JHipsterConstants;
+
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
