@@ -10,7 +10,7 @@ public class ClientForwardController {
      * Forwards any unmapped paths (except those containing a period) to the client {@code index.html}.
      * @return forward to client {@code index.html}.
      */
-    @GetMapping(value = "/**/{path:^(?!websocket)[^\\.]*}")
+    @GetMapping(value = {"/{path:[^\\.]*}", "/{path:^(?!websocket).*}/**/{path:[^\\.]*}"})
     public String forward() {
         return "forward:/";
     }
